warning fix part 1

+++ src/LRUCache.cpp
@@ -82,7 +82,7 @@ size_t LRUCache::findSlot(const std::string& track_id) const {
 size_t LRUCache::findLRUSlot() const {
     size_t curr = max_size;
     uint64_t min = UINT64_MAX;
-    for (int i = 0; i<max_size;i++){
+    for (size_t i = 0; i<max_size;i++){
         if (slots[i].isOccupied()) {
             if (min>slots[i].getLastAccessTime()){
                 min = slots[i].getLastAccessTime();
